pyproject.toml
[tool.poetry]
name = "beast"
version = "0.1.0"
description = ""
authors = ["Chris Proctor <chris@chrisproctor.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
retro-games = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
